requirements.txt
Django>=3.11.0
djangorestframework>=3.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
